{"schema":"decodex/commit/1","summary":"Split search scoring modules","authority":"manual"}
